Remove Google Drive 1 hour error mapping
Years ago, there was an issue where Google Drive would always return a
generic "Unable to play the video at this time" error if the video was 1
hour long or longer.  This is no longer needed, and results in a
confusing error message when the video is unavailable for other reasons.